Add salary automatic crediting feature with email notification and user balance update

+++ moneymind/database/migrations/0001_01_01_000000_create_users_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('password');
             $table->enum('role', ['user', 'admin'])->default('user');
             $table->decimal('salary', 10, 2)->nullable(); // monthly salary
-            $table->unsignedTinyInteger('credit_day')->nullable(); // day of the month for credit
+            $table->unsignedInteger('credit_day')->nullable(); // day of the month for credit
             $table->timestamp('last_login_at')->nullable();
             $table->rememberToken();
             $table->timestamps();
